Reduce the number of irrelevant exceptions in build trace display
Exceptions which are transitively caused by other tasks failing are not
really needed for displaying in the build trace, as the root cause is
that first task failing. The goal is similar how the exceptions are
displayed in the console, only ones that are root causes should be
displayed.
Similar to this, tasks that just fail this way, shouldn't be displayed
on the timeline either. They are just visual clutter.

This may need some fine tuning in the future, based on experience.